Remove default compartment
Since the compartment volume was 1, removing it does nothing, and makes it more clear that the default is being used.  Also fixes the units of the kinetic laws, making all the k's have the same units as they should.

Don't worry about declaring species; if they show up in a reaction, Antimony flags them as species anyway.